Implement filter
Now the locations in the list can be filtered by keyword.

If the location display name doesn’t match the search term, then the
location will be hidden from the list and it’s marker will be removed
from the map.

Filtering will be triggered when the text in the filter input field
changes