Change: hardcoded place name to our location json
Google Maps limited number of geocoding request per second. We can only
have around 5 requests per second. Using Google Maps to translate the
name from coordinate is not feasible, as a result, I decided to
hardcode the place name into the JSON file instead.

Because we have the name from JSON, so geocoder is not needed anymore.
On the order hand, the name is immutable, so we don’t need to change it
on the fly, and can use it to search for Wikipedia article.