Refactor: add marker flow
In past, we add marker to the map before Google Maps give us the info
of the place, so some of the data of the place might be missing in
marker.

Now we add marker to the map after the place info from Google Maps are
ready. We assume that the last info we get from Google Maps is the
place id. So we listen the the changes in marker id and add the marker
to the map when the id is ready.