Fix: potentially too many attempts to query place info from Google Map
We subscribe to any changes in places array in the view model. And
pushing each element to the array will result in fetching info of all
elements in the array from Google Maps. Although we do check the query
is finished or not before querying, however, the querying of the place
may be ongoing, so we may be querying the same place at the same time.

In order to solve this issue, we store the places in standard array,
and only do update all info after the array is filled.